Fix Error Causing Duplicate Responses When Token Expired
If the access token expired after the first page of a multiple
page response had been retreived, the previous code would
refresh the access token, but start the request over again from
page 1.  This would lead to duplicate data appearing in the
response.

It was a rare bug that would require exact timing to occur.

This code stores the page request token, and reuses it after the
access token is refreshed to avoid any duplicate data.